+++ src/io/Export.hpp
@@ -11,21 +11,10 @@ namespace precice {
 namespace precice {
 namespace io {
 
-/**
- * @brief Abstract base class of all classes exporting container data structures.
- */
+/// Abstract base class of all classes exporting container data structures.
 class Export
 {
 public:
-
-  /**
-   * @brief Constructor.
-   *
-   * @param[in] location  Location of export, filepath, e.g.
-   * @param[in] name Name of the export, filename, e.g.
-   * @param[in] exportable Vistable/s to be exported.
-   */
-  //Export();
 
   virtual ~Export() {}
 
@@ -36,13 +25,13 @@ public:
    * @brief Does export. Has to be implemented in subclass.
    *
    * @param[in] name Filename (without path).
-   * @param[in] name Location (path without filename).
+   * @param[in] location Location (path without filename).
    * @param[in] mesh Mesh to be exported.
    */
   virtual void doExport (
     const std::string& name,
     const std::string& location,
-    mesh::Mesh&        mesh ) =0;
+    mesh::Mesh & mesh) = 0;
 };
 
 }} // namespace precice, io