+++ src/io/config/ExportConfiguration.cpp
@@ -18,7 +18,7 @@ ExportConfiguration:: ExportConfiguration(xml::XMLTag& parent)
     tags.push_back(tag);
   }
 
- auto attrLocation = XMLAttribute<std::string>(ATTR_LOCATION, "")
+  auto attrLocation = XMLAttribute<std::string>(ATTR_LOCATION, "")
       .setDocumentation("Directory to export the files to.");
 
   auto attrTimestepInterval = makeXMLAttribute(ATTR_TIMESTEP_INTERVAL, 1)